--- src/MapEvent.js
@@ -1,11 +1,0 @@
-import { useMapEvents } from 'react-leaflet';
-
-const MapEvent = ({ handleMapClick }) => {
-  useMapEvents({
-    click: handleMapClick,
-  });
-
-  return null;
-};
-
-export default MapEvent;
